let it be

+++ src/main/java/lotr/client/render/tileentity/LOTRRenderUtumnoPortal.java
@@ -22,7 +22,7 @@ public class LOTRRenderUtumnoPortal extends TileEntitySpecialRenderer {
 			GL11.glTranslatef((float) d + 0.5f, (float) d1 + 1.0f + i * 0.5f, (float) d2 + 0.5f);
 			Tessellator tessellator = Tessellator.instance;
 			tessellator.startDrawingQuads();
-			tessellator.setColorRGBA_F(0.0f, 0.0f, 0.0f, (float) (passes - i) / (float) passes);
+			tessellator.setColorRGBA_F(0.0f, 0.0f, 0.0f, (float) (passes - i) / passes);
 			double width = 0.5;
 			tessellator.addVertexWithUV(width, 0.0, width, 0.0, 0.0);
 			tessellator.addVertexWithUV(width, 0.0, -width, 0.0, 0.0);
