+++ src/main/java/lotr/common/entity/npc/LOTRTradeEntry.java
@@ -32,7 +32,7 @@ public class LOTRTradeEntry {
 
 	public float getLockedProgress() {
 		if (theTrader != null && theTrader.shouldLockTrades()) {
-			return (float) recentTradeValue / (float) theTrader.getLockTradeAtValue();
+			return (float) recentTradeValue / theTrader.getLockTradeAtValue();
 		}
 		return 0.0f;
 	}
